[autopatch] revisited API to match recent Python implementation
