+++ Rakefile
@@ -7,7 +7,7 @@ require 'appraisal'
 Rake::TestTask.new(:test) do |t|
   t.libs << %w(test lib)
   t.test_files = FileList['test/**/*_test.rb'].reject do |path|
-    path.include?('contrib')
+    path.include?('contrib') || path.include?('monkey_test.rb')
   end
 end
 
@@ -19,13 +19,13 @@ end
 Rake::TestTask.new(:contrib) do |t|
   t.libs << %w(test lib)
   t.test_files = FileList['test/contrib/**/*_test.rb'].reject do |path|
-    path.include?('rails') || path.include?('autopatch_test.rb')
+    path.include?('rails') || path.include?('monkey_test.rb')
   end
 end
 
-Rake::TestTask.new(:autopatch) do |t|
+Rake::TestTask.new(:monkey) do |t|
   t.libs << %w(test lib)
-  t.test_files = FileList['test/contrib/autopatch_test.rb']
+  t.test_files = FileList['test/monkey_test.rb']
 end
 
 Rake::TestTask.new(:benchmark) do |t|
