+++ circle.yml
@@ -45,7 +45,7 @@ test:
     - rvm $RAILS5_VERSIONS --verbose do appraisal rails5-postgres rake rails
     - rvm $RAILS5_VERSIONS --verbose do appraisal rails5-mysql2 rake rails
     - rvm $MRI_VERSIONS,$JRUBY_VERSIONS --verbose do appraisal contrib rake contrib
-    - rvm $MRI_VERSIONS,$JRUBY_VERSIONS --verbose do appraisal contrib rake autopatch
+    - rvm $MRI_VERSIONS,$JRUBY_VERSIONS --verbose do appraisal contrib rake monkey
     - rvm 2.3.1 --verbose do rake benchmark
     - rvm jruby-9.1.5.0 --verbose do rake benchmark
